# api/v1/swagger.yaml
...
swagger: "2.0"
...
info:
  version: 1.0.0
  title: Refocus API
  # description: >-
  #   A short description of the application. GFM syntax can be used for rich text
  #   representation.
  # termsOfService: The Terms of Service for the API.
  # contact:
  #   name: The identifying name of the contact person/organization.
  #   url: http://focus.salesforce.com/contact
  #   email: focusapi@salesforce.com
  # license:
  #   name: The license name used for the API.
  #   url: http://focus.salesforce.com/license

# =============================================================================
host: localhost:10010
...
basePath: /v1
...
schemes:
  - https
  - http

# =============================================================================
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
# =============================================================================
consumes:
  - application/json

# =============================================================================
produces:
  - application/json

# =============================================================================
paths:
...
  /lenses:
    x-swagger-router-controller: lenses
    get:
      security:
        - jwt: []
      summary: Find a lens or lenses
      tags: [ lenses ]
      description: >-
        Find a lens or lenses. You may query using field filters with asterisk
        (*) wildcards. You may also optionally specify sort, limit, offset, and
        a list of fields to include in the response. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      externalDocs:
        description: >-
          GET /lenses
        url: http://focus.salesforce.com/docs/api/lenses
      operationId: findLenses
      parameters:
        -
          $ref: "#/parameters/LensesFieldsParam"
        - name: sort
          in: query
          description: >
            Specify the sort order using a field name, e.g. '...?sort=name'. Prepend the
            field name with a minus sign to specify descending order, e.g.
            '...?sort=-name'.
          required: false
          type: array
          collectionFormat: csv
          items:
            type: string
            enum:
              - description
              - helpEmail
              - helpUrl
              - id
              - isCustom
              - isPublished
              - name
              - sourceDescription
              - sourceName
              - sourceVersion
              - installedBy
              - createdAt
              - updatedAt
              - -description
              - -helpEmail
              - -helpUrl
              - -id
              - -isCustom
              - -isPublished
              - -name
              - -sourceDescription
              - -sourceName
              - -sourceVersion
              - -installedBy
              - -createdAt
              - -updatedAt
        -
          $ref: "#/parameters/limitParam"
        -
          $ref: "#/parameters/offsetParam"
        -
          name: description
          in: query
          required: false
          type: string
          description: >-
            A description of the lens.
        -
          name: helpEmail
          in: query
          required: false
          type: string
          description: >-
            The email address where a user can go to get more help about the
            lens. A lens may choose to display this information to the user e.g.
            if the user clicks a “help” icon attached to the lens.
        -
          name: helpUrl
          in: query
          required: false
          type: string
          description: >-
            The url where a user can go to get more help about the lens. A lens
            may choose to display this information to the user e.g. if the user
            clicks a “help” icon attached to the lens.
        -
          name: isCustom
          type: boolean
          in: query
          required: false
          description: >-
            True for any lens which is installed by an administrator (rather
            than pre-installed in Focus).
        -
          name: isPublished
          type: boolean
          in: query
          required: false
          description: >-
            True for any lens which is available to users.
        -
          name: name
          type: string
          in: query
          required: false
          description: The name of the lens.
        -
          name: sourceDescription
          in: query
          required: false
          type: string
          description: >-
            The original lens publisher's description of the lens.
        -
          name: sourceName
          type: string
          in: query
          required: false
          description: The original lens publisher's name of the lens.
        -
          name: sourceVersion
          type: string
          in: query
          required: false
          description: The original lens publisher's version of the lens.
        -
          name: version
          type: string
          in: query
          required: false
          description: The version of the lens.
      responses:
        200:
          description: >-
            Success, returns a list of lenses.
          schema:
            type: array
            items:
              $ref: "#/definitions/LensesResponse"
        400:
          $ref: "#/responses/400"
        401:
          $ref: "#/responses/401"
        default:
          $ref: "#/responses/genericError"
    post:
      security:
        - jwt: []
      summary: Install a new lens
      tags: [ lenses ]
      description: Install a new lens. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: postLens
      consumes:
        - multipart/form-data
      parameters:
        -
          name: description
          in: formData
          description: >-
            A description of the lens. Leave empty if you want to use the
            description provided by the lens publisher (i.e. sourceDescription).
          type: string
        -
          name: isPublished
          in: formData
          description: >-
            Default false. When set to true, the lens is available to all users.
          type: boolean
          default: false
        -
          name: library
          description: >-
            An archive (zip) of metadata and resources required for a browser to
            render the lens.
          in: formData
          type: file
          format: binary
          required: true
        -
          name: name
          in: formData
          description: >-
            A unique name for the lens. Leave empty if you want to use the
            name provided by the lens publisher (i.e. sourceName).
          type: string
        -
          name: version
          in: formData
          description: >-
            A version for the lens. Leave empty if you want to use the
            version provided by the lens publisher (i.e. sourceVersion).
          type: string
      responses:
        201:
          description: Returns created lens
          schema:
            $ref: "#/definitions/LensesResponse"
        400:
          $ref: "#/responses/400"
        401:
          $ref: "#/responses/401"
        403:
          $ref: "#/responses/403"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
  /lenses/{key}:
    x-swagger-router-controller: lenses
    delete:
      security:
        - jwt: []
      summary: Uninstall the specified lens
      tags: [ lenses ]
      description: >-
        Uninstall the specified lens. You will not be allowed to uninstall the
        lens if it is used by any perspectives. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: deleteLens
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the lens to delete.
          required: true
          type: string
      responses:
        200:
          description: >-
            Success, returns the deleted lens.
          schema:
            $ref: "#/definitions/LensesResponse"
        400:
          $ref: "#/responses/400"
        401:
          $ref: "#/responses/401"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    get:
      security:
        - jwt: []
      summary: Retrieve the specified lens
      tags: [ lenses ]
      description: >-
        Retrieve the specified lens by its id or name. You may also optionally
        specify a list of fields to include in the response. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: getLens
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the lens to retrieve.
          required: true
          type: string
        -
          $ref: "#/parameters/LensesFieldsParam"
      responses:
        200:
          description: >-
            Success, returns the specified lens.
          schema:
            $ref: "#/definitions/LensesResponse"
        400:
          $ref: "#/responses/400"
        401:
          $ref: "#/responses/401"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    patch:
      security:
        - jwt: []
      summary: Update selected metadata fields for the specified lens
      tags: [ lenses ]
      description: >-
        Update selected metadata fields for the specified lens. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      operationId: patchLens
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the lens to update.
          required: true
          type: string
        -
          name: queryBody
          in: body
          description: Request body
          required: true
          schema:
            type: object
            description: The lens to update.
            properties:
              description:
                description: >-
                  A description of the lens. Leave empty if you want to use the
                  description provided by the lens publisher (i.e. sourceDescription).
                type: string
              isPublished:
                description: >-
                  Default false. When set to true, the lens is available to all users.
                type: boolean
                default: false
              name:
                description: >-
                  A unique name for the lens. Leave empty if you want to use the
                  name provided by the lens publisher (i.e. sourceName).
                type: string
              version:
                description: >-
                  A version for the lens. Leave empty if you want to use the
                  version provided by the lens publisher (i.e. sourceVersion).
                type: string
      responses:
        200:
          description: >-
            Success, returns the updated lens.
          schema:
            $ref: "#/definitions/LensesResponse"
        400:
          $ref: "#/responses/400"
        401:
          $ref: "#/responses/401"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"
    put:
      security:
        - jwt: []
      summary: Install a new version of the specified lens
      tags: [ lenses ]
      description: >-
        Install a new version of the specified lens. If the Refocus configuration parameter `useAccessToken` is set to `true`, you must include an `Authorization` request header with your [JSON Web Token](https://tools.ietf.org/html/rfc7519) (JWT) as the value. You can get a token using `POST /v1/register` or `POST /v1/token`.
      consumes:
        - multipart/form-data
      operationId: putLens
      parameters:
        -
          name: key
          in: path
          description: >-
            The id or name of the lens to update.
          required: true
          type: string
        -
          name: library
          description: >-
            An archive (zip) of metadata and resources required for a browser to
            render the lens.
          in: formData
          type: file
          required: true
        -
          name: description
          in: formData
          description: >-
            A description of the lens. Leave empty if you want to use the
            description provided by the lens publisher (i.e. sourceDescription).
          type: string
        -
          name: isPublished
          in: formData
          description: >-
            Default false. When set to true, the lens is available to all users.
          type: boolean
          default: false
        -
          name: name
          in: formData
          description: >-
            A unique name for the lens. Leave empty if you want to use the
            name provided by the lens publisher (i.e. sourceName).
          type: string
        -
          name: version
          in: formData
          description: >-
            A version for the lens. Leave empty if you want to use the
            version provided by the lens publisher (i.e. sourceVersion).
          type: string
      responses:
        200:
          description: >-
            Success, returns the updated lens.
          schema:
            $ref: "#/definitions/LensesResponse"
        400:
          $ref: "#/responses/400"
        401:
          $ref: "#/responses/401"
        403:
          $ref: "#/responses/403"
        404:
          $ref: "#/responses/404"
        default:
          $ref: "#/responses/genericError"

  # ---------------------------------------------------------------------------
...
definitions:
  AuthenticationError:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  BadRequest:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  ErrorResponse:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors
  #  required:
  #    - message
  HATEOAS:
    type: object
    description: >
      Hypertext As The Engine Of Application State.
    properties:
      href:
        type: string
        maxLength: 2082
        readOnly: true
        description: >
          URL of the related HATEOAS link you can use for subsequent calls.
      rel:
        type: string
        maxLength: 255
        readOnly: true
        description: >
          Link relation that describes how this link relates to the previous call.
      method:
        type: string
        readOnly: true
        maxLength: 255
        description: >
          The HTTP method required for the related call.
    required:
      - href
      - rel
      - method
...
  ResourceNotFoundError:
    properties:
      errors:
        type: array
        items:
          type: object
    required:
      - errors

    # properties:
    #   message:
    #     type: string
    #   resourceType:
    #     type: string
    #     enum:
    #       - Aspect
    #       - Lens
    #       - Perspective
    #       - Profile
    #       - Sample
    #       - Subject
    #       - User
    #   resourceKey:
    #     type: string
    # required:
    #   - message
...
  LensesResponse:
    type: object
    description: >
      A user interface component which renders subject/sample data with a
      particular visual representation.
    properties:
      description:
        type: string
        readOnly: true
        maxLength: 4096
        description: >
          A description of the lens.
      helpEmail:
        type: string
        readOnly: true
        description: >
          The email address where a user can go to get more help about the lens.
          A lens may choose to display this information to the user e.g. if the
          user clicks a "help" icon attached to the lens.
        maxLength: 254
      helpUrl:
        type: string
        readOnly: true
        maxLength: 2082
        description: >
          The url where a user can go to get more help about the lens. A lens may
          choose to display this information to the user e.g. if the user clicks a
          "help" icon attached to the lens.
      id:
        type: string
        readOnly: true
        description: >
          The lens id.
      isCustom:
        type: boolean
        readOnly: true
        default: true
        description: >
          True for any lens which is installed by an administrator (rather than
            pre-installed in Focus).
      isPublished:
        type: boolean
        readOnly: true
        description: >
          True when the lens is available to users.
      name:
        type: string
        maxLength: 60
        readOnly: true
        pattern: ^[0-9A-Za-z_\\-]{0,60}$
        description: >
          The name of the lens.
      sourceDescription:
        type: string
        readOnly: true
        maxLength: 4096
        description: >
          The original lens publisher's description of the lens.
      sourceName:
        type: string
        maxLength: 60
        readOnly: true
        pattern: ^[0-9A-Za-z_\\-]{0,60}$
        description: >
          The original lens publisher's name of the lens.
      sourceVersion:
        type: string
        readOnly: true
        description: >
          The original lens publisher's version of the lens.
      thumbnailUrl:
        type: string
        maxLength: 2082
        readOnly: true
        description: >
          Optional thumbnail image to represent the lens.
      installedBy:
        readOnly: true
        type: string
        description: >
          TODO
      createdAt:
        readOnly: true
        type: string
        description: >
          TODO
      updatedAt:
        readOnly: true
        type: string
        description: >
          TODO
      apiLinks:
        readOnly: true
        type: array
        items:
          $ref: "#/definitions/HATEOAS"
        description: >
          Hypertext As The Engine Of Application State.
...
parameters:
  limitParam:
    name: limit
    in: query
    description: >
        Maximum number of records to return in the response; also useful for
        pagination in the UI.
    required: false
    type: integer
  offsetParam:
    name: offset
    in: query
    description: >
        Offset for paginating the response.
    required: false
    type: integer
...
  LensesFieldsParam:
    name: fields
    in: query
    description: >
      Comma-delimited list of field names to include in the response.
    required: false
    type: array
    collectionFormat: csv
    items:
      type: string
      enum:
        - description
        - helpEmail
        - helpUrl
        - id
        - isCustom
        - name
        - thumbnailUrl
        - installedBy
        - createdAt
        - updatedAt
...
responses:
  400:
    description: Bad Request
    schema:
      $ref: '#/definitions/BadRequest'
  401:
    description: Authentication Error
    schema:
      $ref: '#/definitions/AuthenticationError'
  403:
    description: Forbidden
    schema:
      $ref: '#/definitions/ErrorResponse'
  404:
    description: Resource Not Found
    schema:
      $ref: '#/definitions/ResourceNotFoundError'
  genericError:
    description: Error
    schema:
      $ref: '#/definitions/ErrorResponse'

# =============================================================================